rename FAQS.tsx to index.tsx
