+++ frontend/src/components/FAQ/index.tsx
